requirements.txt
flask==2.0.2
flask_login==0.5.0
flask_migrate==3.1.0
WTForms==3.0.1
flask_wtf==1.0.0
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.29
email_validator==1.1.3
gunicorn==20.1.0
jinja2==3.0.3
flask-restx==0.5.1
Werkzeug==2.0.3
python-dotenv==0.19.2
Flask-Minify==0.37
blinker==1.6.2
pyOpenSSL==23.2.0
flask-socketio==5.3.6
pymysql==1.1.0
paho-mqtt==1.6.1
apscheduler==3.10.4
flask-apscheduler==1.12.3
click==8.0.4
itsdangerous==2.0.1
python-dateutil==2.8.2
eventlet==0.33.3
simple-websocket==0.10.1
sqlalchemy-utils==0.41.2